+++ backend/src/main/java/org/tarjetamish/contact/repository/impl/ContactRepository.java
@@ -17,5 +17,5 @@ public interface ContactRepository {
 
     int save(Contact contact);
 
-    int deleteById(int id);
+    int deleteById(Long id);
 }
